MixedModels.likelihoodratiotest — Method

likelihoodratiotest(m)

Calculate likelihoodratiotest

StatsAPI.coefnames — Method

coefnames(term)

coefnames of a TimeExpandedTerm concatenates the basis-function name with the kronecker product of the term name and the basis-function colnames. Separator is ' : ' Some examples for a firbasis: basis313 : (Intercept) : 0.1 basis313 : (Intercept) : 0.2 basis_313 : (Intercept) : 0.3 ...

StatsAPI.fit! — Method

fit!(uf::UnfoldModel,data::Union{<:AbstractArray{T,2},<:AbstractArray{T,3}}) where {T<:Union{Missing, <:Number}}

Fit a DesignMatrix against a 2D/3D Array data along its last dimension Data is typically interpreted as channel x time (with basisfunctions) or channel x time x epoch (for mass univariate)

Returns an UnfoldModel object

Examples

StatsAPI.fit — Method

fit(type::UnfoldModel,f::FormulaTerm,tbl::DataFrame,data::Array{T,3},times) fit(type::UnfoldModel,f::FormulaTerm,tbl::DataFrame,data::Array{T,2},basisfunction::BasisFunction) fit(type::UnfoldModel,d::Dict,tbl::DataFrame,data::Array)

Generates Designmatrix & fits model, either mass-univariate (one model per epoched-timepoint) or time-expanded (modeling linear overlap).

Examples

Mass Univariate Linear

julia> data,evts = loadtestdata("testCase1")
julia> data_r = reshape(data,(1,:))
julia> data_e,times = Unfold.epoch(data=data_r,tbl=evts,τ=(-1.,1.9),sfreq=10) # cut the data into epochs. data_e is now ch x times x epoch

julia> f  = @formula 0~1+continuousA+continuousB # 1
julia> model = fit(UnfoldModel,f,evts,data_e,times)

Timexpanded Univariate Linear

julia> basisfunction = firbasis(τ=(-1,1),sfreq=10,name="A")
julia> model = fit(UnfoldModel,Dict(Any=>(f,basisfunction),evts,data_r)

StatsModels.modelcols — Method

modelcols(term, tbl)

calculates the actual designmatrix for a timeexpandedterm. Multiple dispatch on StatsModels.modelcols

StatsModels.modelcols — Method

modelcols(term, tbl)

This function timeexpands the random effects and generates a ReMat object

Unfold.LinearMixedModel_wrapper — Method

LinearMixedModel_wrapper(form, data, Xs; wts)

Wrapper to generate a LinearMixedModel. Code taken from MixedModels.jl and slightly adapted.

Unfold.apply_basisfunction — Method

apply_basisfunction(form, basisfunction, eventfields)

timeexpand the rhs-term of the formula with the basisfunction

Unfold.combineDesignmatrices — Method

combineDesignmatrices(X1, X2)

Combine two UnfoldDesignmatrices. This allows combination of multiple events.

This also allows to define events with different lengths.

Not supported for models without timebasis, as it is not needed there (one can simply run multiple models)

Examples

julia>  basisfunction1 = firbasis(τ=(0,1),sfreq = 10,name="basis1")
julia>  basisfunction2 = firbasis(τ=(0,0.5),sfreq = 10,name="basis2")
julia>  Xdc1          = designmatrix(UnfoldLinearModelContinuousTime(Dict(Any=>(@formula 0~1,basisfunction1)),tbl_1)
julia>  Xdc2          = designmatrix(UnfoldLinearModelContinuousTime(Dict(Any=>(@formula 0~1,basisfunction2)),tbl_2)
julia>  Xdc = Xdc1+Xdc2 

Unfold.designmatrix — Method

designmatrix(
    type,
    f,
    tbl,
    basisfunction;
    contrasts,
    kwargs...
)

designmatrix(type, f, tbl; kwargs...) Return a DesignMatrix used to fit the models.

Arguments

type::UnfoldModel

f::FormulaTerm: Formula to be used in this designmatrix

tbl: Events (usually a data frame) to be modelled

basisfunction::BasisFunction: basisfunction to be used in modeling (if specified)

contrasts::Dict: (optional) contrast to be applied to formula

eventfields::Array: (optional) Array of symbols which are passed to basisfunction event-wise.

First field of array always defines eventonset in samples. Default is [:latency]

Examples

julia>  designmatrix(UnfoldLinearModelContinuousTime,Dict(Any=>(f,basisfunction1),tbl)

Unfold.fake_lmm — Method

fake_lmm(m, k)

Returns a partial LMM model (non-functional due to lacking data) to be used in likelihoodratiotests. k to selcet which of the modelfit's to fake

Unfold.firbasis — Method

firbasis(τ, sfreq, name)

Generate a sparse FIR basis around the τ timevector at sampling rate sfreq. This is useful if you cannot make any assumptions on the shape of the event responses. If unrounded events are supplied, they are split between samples. E.g. event-latency = 1.2 will result in a "0.8" and a "0.2" entry.

Examples

Generate a FIR basis function from -0.1s to 0.3s at 100Hz

julia>  f = firbasis([-0.1,0.3],100)

Evaluate at an event occuring at sample 103.3

julia>  f(103.3)

Unfold.firkernel — Method

firkernel(e, times)

Calculate a sparse firbasis

Examples

julia>  f = firkernel(103.3,range(-0.1,step=0.01,stop=0.31))

Unfold.hrfbasis — Method

hrfbasis(TR; parameters, name)

Generate a Hemodynamic-Response-Functio (HRF) basis with inverse-samplingrate "TR" (=1/FS)

Optional Parameters p: defaults {seconds} p(1) - delay of response (relative to onset) 6 p(2) - delay of undershoot (relative to onset) 16 p(3) - dispersion of response 1 p(4) - dispersion of undershoot 1 p(5) - ratio of response to undershoot 6 p(6) - onset {seconds} 0 p(7) - length of kernel {seconds} 32

Examples

Generate a HRF basis function object with Sampling rate 1/TR. And evaluate it at an event occuring at TR 103.3 with duration of 4.1 TRs

julia>  f = hrfbasis(2.3)
julia>  f(103.3,4.1)

Unfold.hrfkernel — Method

hrfkernel(e, TR, p)

Calculate a HRF kernel. Input e can be [onset duration]

Examples

julia>  f = hrfkernel(103.3,2.3,[6. 16. 1. 1. 6. 0. 32.])

Unfold.linearize — Method

linearize(x)

Flatten a 1D array from of a 2D/3D array. Also drops the empty dimension

Unfold.pvalues — Method

pvalues(lrtvec)

Unfold-Method: return pvalues of likelihoodratiotests, typically calculated:

Examples

julia> pvalues(likelihoodratiotest(m1,m2))

where m1/m2 are UnfoldLinearMixedModel's

Tipp: if you only compare two models you can easily get a vector of p-values:

julia> vcat(pvalues(likelihoodratiotest(m1,m2))...)

Multiple channels are returned linearized at the moment, as we do not have access to the amount of channels after the LRT, you can do:

julia> reshape(vcat(pvalues(likelihoodratiotest(m1,m2))...),ntimes,nchan)'

Unfold.time_expand — Method

time_expand(X, term, tbl)

performs the actual time-expansion in a sparse way.

Get the non-timeexpanded designmatrix X from StatsModels.

evaluate the basisfunction kernel at each event

calculate the necessary rows, cols and values for the sparse matrix

Returns SparseMatrixCSC

Unfold.time_expand_getRandomGrouping — Method

time_expand_getRandomGrouping(
    tblGroup,
    tblLatencies,
    basisfunction
)

Get the timeranges where the random grouping variable was applied

Unfold.zeropad — Method

zeropad(X, data)

Equates the length of data and designmatrix by cutting the shorter one

The reason we need this is because when generating the designmatrix, we do not know how long the data actually are. We only assume that event-latencies are synchronized with the data