Unfold.BasisFunction — Type

See FIRBasis for an examples

a BasisFunction should implement:
kernel()
collabel() [default "colname_basis"]
colnames()
times()
name()
shiftOnset() [default 0]

Unfold.FIRBasis — Type

Defines a FIRBasisfunction which can be called for each event, defining the time-expanded basis kernel

struct FIRBasis <: Unfold.BasisFunction

kernel

a design-matrix kernel function used to timeexpand, given a timepoint in 'sample' timeunits

times

vector of times along rows of kernel-output (in seconds)

name

name of the event, random 1:1000 if unspecified

shiftOnset

by how many samples do we need to shift the event onsets? This number is determined by how many 'negative' timepoints the basisfunction defines

Examples

julia>  b = BasisFunction(kernelfunction,"derivative",["f(x)"],range(0,(length(kernelfunction([0, 1]))-1)*TR,step=TR),"hrf_kernel","basis_A",0)

Unfold.TimeExpandedTerm — Type

Object with a term and an applicable BasisFunction and a eventfield that are later passed to the basisfunction.

struct TimeExpandedTerm{T<:StatsModels.AbstractTerm} <: StatsModels.AbstractTerm

term

Term that the basis function is applied to. This is regularly called in other functions to get e.g. term-coefnames and timeexpand those

basisfunction

Kernel that determines what should happen to the designmatrix of the term

eventfields

Which fields of the event-table should be passed to the basisfunction.Important: The first entry has to be the event-latency in samples!

Examples

julia>  b = TimeExpandedTerm(term,kernel,[:latencyTR,:durationTR])

Unfold.ZeroCorr2 — Type

To be explored abstractTerm structure to automatically detect randomeffects. Not in use